# Thread LeafManager and NodeManager construction