The main motivation here was to eliminate the feature where the leaf nodes of the LeafManager could be used in the NodeManager construction to improve performance (as it was causing complexity in attempting to re-use the NodeList elsewhere). The simplest solution to achieving this was to just thread the construction of the LeafManager and the NodeManager. 

I also converted the NodeList and the LeafManager over to use unique ptr arrays instead of std::deque and raw ptr arrays (I expect these classes were written before C++11 features were available). Note that I've treated the NodeList as internal API due to it being described as such in the documentation, so made more invasive changes than I would otherwise do - it's non-trivial to try and also support backwards compatibility in this class. Finally, I've marked LeafManager::getNodes() methods as deprecated, they were only used in NodeManager construction to match the methods found in the tree hierarchy, so feels safe to mark them to be removed if we remove this code path. 

Here's construction times using the old implementation based on a 2.3 billion voxel sphere:

* LeafManager construction - 127ms
* NodeManager construction from Tree - 161ms
* NodeManager construction from LeafManager - 81ms

And this new threaded implementation:

* LeafManager construction - 18ms
* NodeManager construction from Tree - 19ms
* NodeManager construction from LeafManager - 20ms (now uses the same implementation as construction from Tree)

I also tested LeafManager construction using the data sets in the testActiveLeafVoxelCount to demonstrate scalability of this implementation down to cases with fewer leaf nodes and more unbalanced hierarchies:

* Old Method 1 - 0.03ms
* Old Method 2 - 8.7ms
* Old Method 3 - 12.0ms
* Old Method 4 - 68.0ms

* New Method 1 - 0.04ms
* New Method 2 - 3.7ms
* New Method 3 - 2.8ms
* New Method 4 - 12.8ms