@danrbailey 2020-07-24
> Haven't reviewed this fully but, due to the nature of the changes, I propose these don't go in to 7.1 (I'd rather not re-live [OVDB-117](https://jira.aswf.io/browse/OVDB-117))

Yeah I don't mind not including this for 7.1. Not going to introduce it right now, but might be worth having a new remote branch for this work while we're waiting on 7.1 to go out.