@danrbailey 2020-08-14
Closing this PR, I've rolled these changes into #785.